cicero-app: criação Create e Response DTO de Occupation, Major Event e Generation e alteração do tipo de PK em seus Reposiory

+++ cicero-app/src/main/java/com/duende/cicero_app/repository/origin/GenerationRepository.java
@@ -3,4 +3,4 @@ package com.duende.cicero_app.repository.origin;
 import com.duende.cicero_app.model.origin.GenerationModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GenerationRepository extends JpaRepository <GenerationModel, Long> {}
+public interface GenerationRepository extends JpaRepository <GenerationModel, Short> {}
