+++ cicero-app/src/main/java/com/duende/cicero_app/repository/origin/MajorEventRepository.java
@@ -3,4 +3,4 @@ package com.duende.cicero_app.repository.origin;
 import com.duende.cicero_app.model.origin.MajorEventModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MajorEventRepository extends JpaRepository <MajorEventModel, Long> {}
+public interface MajorEventRepository extends JpaRepository <MajorEventModel, Short> {}
