+++ cicero-app/src/main/java/com/duende/cicero_app/repository/origin/OccupationRepository.java
@@ -3,4 +3,4 @@ package com.duende.cicero_app.repository.origin;
 import com.duende.cicero_app.model.origin.OccupationModel;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface OccupationRepository extends JpaRepository <OccupationModel, Long> {}
+public interface OccupationRepository extends JpaRepository <OccupationModel, Short> {}
